requirements.txt
pillow == 11.0.0
pygame == 2.6.1
